requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-cov==4.1.0
allure-pytest==2.13.2
pyexcel-ods==0.6.0
